+++ components/edit/EditShell/EditShell.tsx
@@ -1,11 +1,13 @@
 'use client';
 
-import { motion } from 'motion/react';
-import type { ReactNode } from 'react';
-import type { SceneEditorSurface } from '@/lib/edit/scene-editor-surface';
+import { motion, useReducedMotion } from 'motion/react';
+import { useLayoutEffect, useRef, useState, type ReactNode } from 'react';
+import type { SceneEditorSurface, SurfaceState } from '@/lib/edit/scene-editor-surface';
 import { sceneEditorRegistry } from '@/lib/edit/scene-editor-registry';
-import { useI18n } from '@/lib/hooks/use-i18n';
+import { NOOP_SURFACE } from '@/lib/edit/noop-surface';
 import type { Scene } from '@/lib/types/stage';
+import { CHROME_DURATION, CHROME_EASE, CHROME_STAGGER } from '@/lib/edit/transitions';
+import { StageGrid } from '@/components/edit/StageGrid';
 import { CommandBar } from './CommandBar';
 import { FloatingToolbar } from './FloatingToolbar';
 import { HintRail } from './HintRail';
@@ -13,15 +15,23 @@ import { HintRail } from './HintRail';
 interface EditShellProps {
   readonly scene: Scene;
   /**
-   * Optional left-side navigator slot. v0 ships with this empty — a future
-   * sub-PR will plug in the redesigned slide-navigation surface here. The
-   * prop is preserved as an extension point so Stage doesn't need to grow a
-   * new layout when that lands.
+   * Optional left-side navigator slot. In v0 this is the SlideNavRail
+   * passed from Stage when mode === 'edit'. Surface code never imports
+   * the rail (the prop is the only handoff seam), keeping chrome and
+   * surface separable.
    */
   readonly leftRail?: ReactNode;
+  /**
+   * Right-edge slot of the CommandBar — Stage uses this to hand in the
+   * global controls (settings pill + Pro Switch) when the Stage Header
+   * is hidden, so the entire top chrome reduces to a single bar.
+   */
+  readonly commandTrailing?: ReactNode;
 }
 
-const CHROME_TRANSITION = { duration: 0.28, ease: [0.22, 1, 0.36, 1] as const };
+const CHROME_TRANSITION = { duration: CHROME_DURATION, ease: CHROME_EASE } as const;
+const COMMANDBAR_DELAY = CHROME_STAGGER;
+const LEFT_RAIL_DELAY = CHROME_STAGGER * 2;
 
 /**
  * Pro mode (edit) chrome — mounts inside the canvas slot of Stage, replacing
@@ -39,64 +49,114 @@ const CHROME_TRANSITION = { duration: 0.28, ease: [0.22, 1, 0.36, 1] as const };
  *   │          │ HintRail (AI, reserved)            │
  *   └──────────┴───────────────────────────────────┘
  *
- * When a surface is registered for `scene.type`, EditShell renders that
- * surface's canvas and reads its useSurfaceState() into the CommandBar /
- * FloatingToolbar / HintRail slots. When none is registered, it falls
- * through to the `edit.unsupportedScene` placeholder — the visible v0
- * behavior since no surfaces ship in this PR.
+ * Mount choreography: CommandBar drops in from top, leftRail slides in
+ * from left after a stagger, content opacity-fades in. All three share
+ * the single `CHROME_*` source in `lib/edit/transitions.ts` so timing
+ * stays consistent with the outer Stage-level cross-fade.
+ *
+ * Architecture: this shell resolves `scene.type` to a registered surface
+ * (or falls back to NOOP_SURFACE for unregistered types) and **never
+ * branches into a different component type**. The same `<Frame>` mounts
+ * across every scene-type change — only the `surface.CanvasComponent`
+ * inside the canvas slot swaps. That guarantees CommandBar and `leftRail`
+ * never remount during scene navigation, removing the chrome flicker that
+ * the previous two-branch design caused (PR3a rearch).
  */
-export function EditShell({ scene, leftRail }: EditShellProps) {
-  const surface = sceneEditorRegistry.resolve(scene.type);
+export function EditShell({ scene, leftRail, commandTrailing }: EditShellProps) {
+  const surface = sceneEditorRegistry.resolve(scene.type) ?? NOOP_SURFACE;
+  // Surface state is published from a child runner (keyed by sceneType so it
+  // remounts when the surface identity changes — that's the boundary at which
+  // rules-of-hooks naturally allows a different hook signature). The chrome
+  // around it stays mounted and consumes state via these props.
+  const [state, setState] = useState<SurfaceState | null>(null);
+  const CanvasComponent = surface.CanvasComponent;
 
-  if (surface) {
-    return <EditShellWithSurface scene={scene} surface={surface} leftRail={leftRail} />;
-  }
-  return <EditShellFallback scene={scene} leftRail={leftRail} />;
+  return (
+    <>
+      {/* `key={scene.type}` is the remount boundary. We can't use
+          `surface.sceneType` here because NOOP_SURFACE deliberately reuses
+          'slide' as a placeholder (the SceneType union is closed and NOOP
+          isn't a real type). The scene's own `type` is the actual signal
+          that the hook signature inside `useSurfaceState` is about to
+          change — so we remount the runner exactly when it does, keeping
+          rules-of-hooks happy across the slide ↔ read-only surface swap
+          while the rest of the chrome stays mounted. */}
+      <SurfaceStateRunner key={scene.type} surface={surface} onChange={setState} />
+      <Frame
+        title={scene.title}
+        leftRail={leftRail}
+        history={state?.history}
+        insertItems={state?.insertItems}
+        commands={state?.commands}
+        trailing={commandTrailing}
+      >
+        <CanvasComponent />
+        {state?.hasSelection && <FloatingToolbar actions={state.floatingActions} />}
+        <HintRail hints={state?.hints} />
+      </Frame>
+    </>
+  );
 }
 
-interface ResolvedShellProps {
-  readonly scene: Scene;
-  readonly leftRail?: ReactNode;
-}
-
-function EditShellWithSurface({
-  scene,
+/**
+ * Hidden runner that owns the surface state hook. `key={surface.sceneType}`
+ * ensures it remounts when the surface itself changes (slide → noop) — the
+ * only point at which the hook call signature can vary. Within a single mount
+ * the hook signature is fixed (the surface object is constant), so React's
+ * rules-of-hooks are respected.
+ *
+ * Renders no DOM; state flows up to the chrome via `onChange`. A custom
+ * shallow comparison gates the publish — surface hooks (e.g. slideSurface's
+ * `useSlideSurfaceState`) return a fresh object literal every render, so naive
+ * reference equality would loop infinitely (every publish causes the parent to
+ * re-render, which re-runs this hook, which yields a new ref, which publishes
+ * again, etc.). We only publish when one of the fields the chrome actually
+ * reads has materially changed.
+ */
+function SurfaceStateRunner({
   surface,
-  leftRail,
-}: ResolvedShellProps & { readonly surface: SceneEditorSurface }) {
-  const { t } = useI18n();
-  const sceneTypeLabel = t(`edit.sceneType.${scene.type}`);
-  const title = t('edit.title', { type: sceneTypeLabel });
+  onChange,
+}: {
+  readonly surface: SceneEditorSurface;
+  readonly onChange: (state: SurfaceState) => void;
+}) {
   const state = surface.useSurfaceState();
-  const Canvas = surface.CanvasComponent;
-
-  return (
-    <Frame
-      title={title}
-      leftRail={leftRail}
-      history={state.history}
-      insertItems={state.insertItems}
-      commands={state.commands}
-    >
-      <Canvas />
-      {state.hasSelection && <FloatingToolbar actions={state.floatingActions} />}
-      <HintRail hints={state.hints} />
-    </Frame>
-  );
+  const lastRef = useRef<SurfaceState | null>(null);
+  useLayoutEffect(() => {
+    if (surfaceStateEqual(state, lastRef.current)) return;
+    lastRef.current = state;
+    onChange(state);
+  });
+  return null;
 }
 
-function EditShellFallback({ scene, leftRail }: ResolvedShellProps) {
-  const { t } = useI18n();
-  const sceneTypeLabel = t(`edit.sceneType.${scene.type}`);
-  const title = t('edit.title', { type: sceneTypeLabel });
-
-  return (
-    <Frame title={title} leftRail={leftRail}>
-      <div className="flex h-full w-full items-center justify-center text-sm text-zinc-500 dark:text-zinc-400">
-        {t('edit.unsupportedScene', { type: sceneTypeLabel })}
-      </div>
-    </Frame>
-  );
+/**
+ * Field-by-field equality for the subset of SurfaceState that the chrome
+ * reads (CommandBar history/insertItems/commands + Frame.hasSelection +
+ * floating/hints). Reference-equal `content` is the canonical signal that
+ * the slide buffer hasn't changed; insert/command arrays are compared by
+ * length + per-item `active` flag (which is what CommandBar styles on).
+ */
+function surfaceStateEqual(a: SurfaceState, b: SurfaceState | null): boolean {
+  if (!b) return false;
+  if (a.content !== b.content) return false;
+  if (a.hasSelection !== b.hasSelection) return false;
+  if ((a.history?.canUndo ?? null) !== (b.history?.canUndo ?? null)) return false;
+  if ((a.history?.canRedo ?? null) !== (b.history?.canRedo ?? null)) return false;
+  if (a.insertItems.length !== b.insertItems.length) return false;
+  for (let i = 0; i < a.insertItems.length; i++) {
+    if (a.insertItems[i].id !== b.insertItems[i].id) return false;
+    if (a.insertItems[i].active !== b.insertItems[i].active) return false;
+    if (a.insertItems[i].disabled !== b.insertItems[i].disabled) return false;
+  }
+  if (a.commands.length !== b.commands.length) return false;
+  for (let i = 0; i < a.commands.length; i++) {
+    if (a.commands[i].id !== b.commands[i].id) return false;
+    if (a.commands[i].disabled !== b.commands[i].disabled) return false;
+  }
+  if (a.floatingActions.length !== b.floatingActions.length) return false;
+  if ((a.hints?.length ?? 0) !== (b.hints?.length ?? 0)) return false;
+  return true;
 }
 
 interface FrameProps {
@@ -105,23 +165,75 @@ interface FrameProps {
   readonly history?: React.ComponentProps<typeof CommandBar>['history'];
   readonly insertItems?: React.ComponentProps<typeof CommandBar>['insertItems'];
   readonly commands?: React.ComponentProps<typeof CommandBar>['commands'];
+  readonly trailing?: ReactNode;
   readonly children: ReactNode;
 }
 
-function Frame({ title, leftRail, history, insertItems, commands, children }: FrameProps) {
+function Frame({
+  title,
+  leftRail,
+  history,
+  insertItems,
+  commands,
+  trailing,
+  children,
+}: FrameProps) {
+  const prefersReducedMotion = useReducedMotion();
+
+  // When reduced motion is requested, drop transforms (y/x) and only fade.
+  // Layout remains static and timing collapses to ~120ms.
+  const cmdInitial = prefersReducedMotion ? { opacity: 0 } : { y: -56, opacity: 0 };
+  const cmdAnimate = prefersReducedMotion ? { opacity: 1 } : { y: 0, opacity: 1 };
+  const railInitial = prefersReducedMotion ? { opacity: 0 } : { x: -32, opacity: 0 };
+  const railAnimate = prefersReducedMotion ? { opacity: 1 } : { x: 0, opacity: 1 };
+
+  const stepTransition = prefersReducedMotion
+    ? { duration: 0.12, ease: CHROME_EASE }
+    : CHROME_TRANSITION;
+
   return (
-    <div className="flex h-full w-full flex-col bg-zinc-50 dark:bg-zinc-950">
-      <motion.div
-        initial={{ y: -56, opacity: 0 }}
-        animate={{ y: 0, opacity: 1 }}
-        transition={CHROME_TRANSITION}
-      >
-        <CommandBar title={title} history={history} insertItems={insertItems} commands={commands} />
-      </motion.div>
-      <div className="flex min-h-0 flex-1">
-        {leftRail}
-        <div className="relative min-h-0 flex-1">{children}</div>
-      </div>
-    </div>
+    <StageGrid
+      className="bg-gradient-to-b from-zinc-100 to-zinc-200 dark:from-zinc-950 dark:to-zinc-900"
+      topSlot={
+        <motion.div
+          initial={cmdInitial}
+          animate={cmdAnimate}
+          transition={{ ...stepTransition, delay: prefersReducedMotion ? 0 : COMMANDBAR_DELAY }}
+        >
+          <CommandBar
+            title={title}
+            history={history}
+            insertItems={insertItems}
+            commands={commands}
+            trailing={trailing}
+          />
+        </motion.div>
+      }
+      leftSlot={
+        leftRail ? (
+          <motion.div
+            initial={railInitial}
+            animate={railAnimate}
+            transition={{ ...stepTransition, delay: prefersReducedMotion ? 0 : LEFT_RAIL_DELAY }}
+            className="h-full shrink-0"
+          >
+            {leftRail}
+          </motion.div>
+        ) : null
+      }
+      centerSlot={
+        // Padded studio frame around the actual scene renderer. Lifted
+        // up from SlideCanvas so the slide and the non-slide read-only
+        // renderers share the exact same canvas bounding rect (no
+        // layout jump when switching scene type). Children render
+        // inside an inner ring/shadow card that the playback
+        // CanvasArea visually mirrors.
+        <div className="relative h-full w-full p-3 sm:p-4">
+          <div className="relative h-full w-full overflow-hidden rounded-xl bg-white ring-1 ring-zinc-200/80 dark:bg-zinc-900 dark:ring-zinc-800/80 shadow-[0_10px_40px_-12px_rgba(15,23,42,0.18)] dark:shadow-[0_10px_40px_-12px_rgba(0,0,0,0.6)]">
+            {children}
+          </div>
+        </div>
+      }
+    />
   );
 }
