+++ components/edit/surfaces/slide/SlideCanvas.tsx
@@ -44,22 +44,16 @@ export function SlideCanvas() {
     return () => document.removeEventListener('keydown', handler);
   }, []);
 
-  // Mark the body while the editor is mounted, so the editor-scoped CSS rule
-  // in globals.css that pins `body.padding-right` to 0 only fires here — not
-  // on non-editor pages where Radix's react-remove-scroll compensation is
-  // still wanted.
-  useEffect(() => {
-    document.body.dataset.maicEditor = 'true';
-    return () => {
-      delete document.body.dataset.maicEditor;
-    };
-  }, []);
-
   return (
     // gestureProps marks pointer-gesture windows so a renderer commit is
     // classified as a real user edit vs ResizeObserver text normalization
-    // (which fires with no gesture in flight).
-    <div className="h-full w-full" {...gestureProps}>
+    // (which fires with no gesture in flight). The padded studio frame
+    // around the canvas now lives in EditShell.Frame so non-slide scenes
+    // (rendered via SceneRenderer in read-only mode) share the exact
+    // same canvas bounding rect — switching scene type no longer
+    // resizes / reflows the frame, which used to cause the slide↔
+    // interactive layout jump.
+    <div className="relative h-full w-full" {...gestureProps}>
       <SceneProvider controller={controller}>
         <Canvas />
       </SceneProvider>
