+++ lib/edit/scene-editor-surface.ts
@@ -96,7 +96,12 @@ export interface SurfaceState<TContent extends SceneContent = SceneContent, TSel
   /** True when the surface considers selection non-empty (drives floating bar). */
   hasSelection: boolean;
 
-  history: SurfaceHistory;
+  /**
+   * Editable surfaces expose undo/redo here. Read-only surfaces (e.g. the
+   * NOOP fallback used for unregistered scene types) omit it; the shell
+   * hides undo/redo controls when undefined.
+   */
+  history?: SurfaceHistory;
 
   insertItems: InsertPaletteItem[];
   floatingActions: FloatingAction[];
